Rename ReactFiberActivityComponent to ReactFiberOffscreenComponent
This is the internal data structure and not the public API.
